+++ src/validation/validation-environment-maker.js
@@ -12,6 +12,7 @@
 exports.init = init;
 
 const fs = require('fs');
+const path = require('path');
 const vm = require('vm');
 const underscore = require('../../lib/underscore/underscore-min');
 const simpleMock = require('../../lib/simple-mock/index');
@@ -22,7 +23,8 @@ function init(docDefinitionsString, originalFilename) {
     displayErrors: true
   };
 
-  const envTemplateString = fs.readFileSync('templates/validation-environment-template.js', 'utf8').trim();
+  const filePath = path.resolve(__dirname, '../../templates/validation-environment-template.js');
+  const envTemplateString = fs.readFileSync(filePath, 'utf8').trim();
 
   // The test helper environment includes a placeholder string called "%DOC_DEFINITIONS_PLACEHOLDER%" that is to be replaced with the
   // contents of the document definitions